move chevron function to scritps
